Move setupTray to appWindow
